% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{data_heatmap}
\alias{data_heatmap}
\title{Decomposition Heatmaps
Plot data block heatmap}
\usage{
data_heatmap(data, show_color_bar = TRUE, title = "", xlab = "", ylab = "")
}
\arguments{
\item{data}{Matrix or data frame. The data to plot as a heatmap.}

\item{show_color_bar}{Boolean.}

\item{title}{Character.}

\item{xlab}{Character.}

\item{ylab}{Character}
}
\value{
A \code{ggplot} object.
}
\description{
Visualization helper for individual data blocks or decomposition components.
Produces a raster heatmap using \pkg{ggplot2} with a rainbow colour scale.
}
