% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw.R
\name{print.jackstraw_rajive}
\alias{print.jackstraw_rajive}
\title{Print method for jackstraw_rajive objects}
\usage{
\method{print}{jackstraw_rajive}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"jackstraw_rajive"}.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Displays a concise summary table of significance results for each block and
joint component.
}
\examples{
\donttest{
n   <- 40; pks <- c(30, 20)
Y   <- ajive.data.sim(K = 2, rankJ = 2, rankA = c(4, 3),
                      n = n, pks = pks, dist.type = 1)
fit <- Rajive(Y$sim_data, c(4, 3))
js  <- jackstraw_rajive(fit, blocks = Y$sim_data, B = 10L)
print(js)
}

}
