% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{wedin_bound_resampling}
\alias{wedin_bound_resampling}
\title{Resampling procedure for the wedin bound}
\usage{
wedin_bound_resampling(
  X,
  signal_basis,
  right_vectors,
  num_samples = 1000,
  num_cores = 2
)
}
\arguments{
\item{X}{Matrix. The data matrix.}

\item{signal_basis}{Matrix. Either U or V signal-space basis after
estimating the signal rank.}

\item{right_vectors}{Boolean. Right multiplication or left multiplication.}

\item{num_samples}{Integer. Number of vectors selected for resampling procedure.}

\item{num_cores}{Integer. Number of parallel cores to use (default 2).}
}
\description{
Resampling procedure for the wedin bound
}
